* Fixed numerous warnings and errors with g++ 3.4.5 * Removed any C++ construct beyond c++0x standard * Updated Makefiles

+++ src/MakespanFormula.cpp
@@ -22,7 +22,7 @@ struct e : public exprtk::ifunction<double>
 		inline double operator()(const double &id, const double &p) 
 		{   
 			Task key((int)id, "key");
-			return tasks.find(key)->getEfficiency(p);
+			return tasks.find(key)->getEfficiency((int)p);
 		}   
 
 	private:
